+++ test/SpagoSpec.hs
@@ -11,7 +11,8 @@ import           Turtle             (ExitCode (..), cd, cp, decodeString, empty,
                                      writeTextFile)
 import           Utils              (checkFileHasInfix, checkFixture, readFixture, runFor,
                                      shouldBeFailure, shouldBeFailureOutput, shouldBeSuccess,
-                                     shouldBeSuccessOutput, shouldBeSuccessOutputWithErr, spago, withCwd)
+                                     shouldBeSuccessOutput, shouldBeSuccessOutputWithErr, spago, withCwd,
+                                     outputShouldEqual)
 
 
 
@@ -582,3 +583,27 @@ spec = around_ setup $ do
       mv "packages.dhall" "packages-old.dhall"
       writeTextFile "packages.dhall" "https://github.com/purescript/package-sets/releases/download/psc-0.13.4-20191025/packages.dhall sha256:f9eb600e5c2a439c3ac9543b1f36590696342baedab2d54ae0aa03c9447ce7d4"
       spago ["list-packages", "--json"] >>= shouldBeSuccessOutput "list-packages.json"
+
+  describe "spago path output" $ do
+    it "Spago should output the correct path" $ do
+
+      -- Create local 'monorepo-1' package that is the real root
+      mkdir "monorepo-1"
+      cd "monorepo-1"
+      spago ["init"] >>= shouldBeSuccess
+
+       -- Create local 'monorepo-2' package that uses packages.dhall on top level
+      mkdir "monorepo-2"
+      cd "monorepo-2"
+      spago ["init"] >>= shouldBeSuccess
+      rm "packages.dhall"
+      writeTextFile "packages.dhall" $ "../packages.dhall"
+      spago ["path", "output"] >>= outputShouldEqual "./../output\n"
+      pure ()
+
+    it "Spago should output the local path when no overrides" $ do
+
+      mkdir "monorepo-1"
+      cd "monorepo-1"
+      spago ["init"] >>= shouldBeSuccess
+      spago ["path", "output", "--no-share-output"] >>= outputShouldEqual "output\n"