+++ CHANGELOG.md
@@ -25,6 +25,7 @@ New features:
 - documentation now includes a step-by-step guide on setting up a Spago/Parcel project (#456)
 - documentation now includes a step-by-step guide on setting up a Spago/Node and Spago/Webpack project (#456-extra)
 - moved warning and error logs to stderr, adjusted logging strategy (#256)
+- `spago path` returns output path so that it can be shared with tools such as `purs-loader`
 
 Bugfixes:
 - Warn (but don't error) when trying to watch missing directories (#406)